Missing jest config for mapped modules
